test/flower/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
dashscope>=1.14.0
serpapi>=0.1.0
requests>=2.28.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0